ai-copilot/requirements.txt
fastapi
uvicorn
streamlit
requests
python-multipart
pydantic
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2
numpy
